# android/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "picweight API",
    "description": "AI-assisted calorie & БЖУ tracker. Photograph what you eat; a server-side estimation agent returns calories and macros.",
    "license": {
      "name": "MIT"
    },
    "version": "master"
  },
  "paths": {
...
    "/api/v1/days/{date}": {
      "get": {
        "tags": [
          "days"
        ],
        "summary": "Get a day's totals and remaining budget",
        "operationId": "get_day",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "description": "Local calendar date, YYYY-MM-DD",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tz_offset",
            "in": "query",
            "description": "Minutes east of UTC to bucket by. Defaults to the offset of the user's\nmost recent meal, then to the profile timezone.",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The day",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayResponse"
                }
              }
            }
          },
          "400": {
            "description": "Unparseable date",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/dishes/recent": {
      "get": {
        "tags": [
          "dishes"
        ],
        "summary": "Recent confirmed dishes",
        "description": "Powers the capture screen's recent-dish chips — one tap sets the dish name, no keyboard. Confirmed meals only.",
        "operationId": "recent_dishes",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "How many chips to return. Defaults to [`DEFAULT_RECENT_LIMIT`], capped at\n[`MAX_RECENT_LIMIT`].",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recent dishes, most recent first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecentDish"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DayResponse": {
        "type": "object",
        "description": "Totals, targets, remaining and the verdict line for one local day.",
        "required": [
          "state",
          "verdict",
          "groups",
          "meals"
        ],
        "properties": {
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupSummary"
            },
            "description": "Sittings, collapsed."
          },
          "meals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MealResponse"
            },
            "description": "Meals not belonging to any sitting, newest first."
          },
          "state": {
            "$ref": "#/components/schemas/DayState",
            "description": "The rules-engine state: consumed, targets, remaining, classification."
          },
          "verdict": {
            "type": "string",
            "description": "One line of plain-language advice, templated or LLM-phrased."
          }
        }
      },
      "DayState": {
        "type": "object",
        "description": "A day's totals against its targets.\n\nCarried verbatim in the `/meals/events` completion payload so a client\nrenders the notification from one message with no follow-up round trip (§9).",
        "required": [
          "date",
          "consumed_kcal",
          "target_kcal",
          "remaining_kcal",
          "consumed_protein_g",
          "target_protein_g",
          "remaining_protein_g",
          "consumed_fat_g",
          "target_fat_g",
          "consumed_carbs_g",
          "target_carbs_g",
          "meals_logged",
          "status"
        ],
        "properties": {
          "consumed_carbs_g": {
            "type": "number",
            "format": "double",
            "description": "Carbohydrate consumed, grams."
          },
          "consumed_fat_g": {
            "type": "number",
            "format": "double",
            "description": "Fat consumed, grams."
          },
          "consumed_kcal": {
            "type": "number",
            "format": "double",
            "description": "Energy consumed, kcal."
          },
          "consumed_protein_g": {
            "type": "number",
            "format": "double",
            "description": "Protein consumed, grams."
          },
          "date": {
            "type": "string",
            "format": "date",
            "description": "The local calendar day these figures cover."
          },
          "meals_logged": {
            "type": "integer",
            "format": "int64",
            "description": "Meals logged on this day (all statuses except `failed`)."
          },
          "remaining_kcal": {
            "type": "number",
            "format": "double",
            "description": "`target_kcal - consumed_kcal`; may be negative."
          },
          "remaining_protein_g": {
            "type": "number",
            "format": "double",
            "description": "Protein still owed, grams; clamped at zero."
          },
          "status": {
            "$ref": "#/components/schemas/DayStatus",
            "description": "Classification derived from the figures above."
          },
          "target_carbs_g": {
            "type": "number",
            "format": "double",
            "description": "Carbohydrate allowance, grams."
          },
          "target_fat_g": {
            "type": "number",
            "format": "double",
            "description": "Fat floor, grams."
          },
          "target_kcal": {
            "type": "number",
            "format": "double",
            "description": "Energy target, kcal. Zero when targets are not computed."
          },
          "target_protein_g": {
            "type": "number",
            "format": "double",
            "description": "Protein floor, grams."
          }
        }
      },
      "DayStatus": {
        "type": "string",
        "description": "How a day is going. Computed from the numbers, never asked of a model.",
        "enum": [
          "on_track",
          "tight",
          "over",
          "protein_unreachable",
          "no_targets"
        ]
      },
      "ErrorBody": {
        "type": "object",
        "description": "The JSON body returned for every [`AppError`].",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Stable machine-readable code, e.g. `not_found`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable detail. Generic for 5xx, specific for 4xx."
          }
        }
      },
...
      "GramsSource": {
        "type": "string",
        "description": "Where an item's gram figure came from.",
        "enum": [
          "Agent",
          "User",
          "Barcode",
          "Recall"
        ]
      },
...
      "GroupSummary": {
        "type": "object",
        "description": "A group as it appears collapsed in the day view.",
        "required": [
          "group_id",
          "member_count",
          "totals",
          "dish_names",
          "eaten_at",
          "has_failures"
        ],
        "properties": {
          "dish_names": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Dish names, for the collapsed row's label."
          },
          "eaten_at": {
            "type": "string",
            "format": "date-time",
            "description": "Earliest `eaten_at` among the members."
          },
          "group_id": {
            "type": "string",
            "description": "Client-generated sitting id."
          },
          "has_failures": {
            "type": "boolean",
            "description": "True when a member failed — the row should say so rather than quietly\nunder-report."
          },
          "member_count": {
            "type": "integer",
            "description": "How many meals it contains.",
            "minimum": 0
          },
          "totals": {
            "$ref": "#/components/schemas/MacroTotals",
            "description": "Combined totals."
          }
        }
      },
...
      "MacroSource": {
        "type": "string",
        "description": "Where an item's macros came from.",
        "enum": [
          "Recall",
          "Model",
          "Barcode",
          "Web",
          "User"
        ]
      },
      "MacroTotals": {
        "type": "object",
        "description": "Summed macros, used for meals, days and groups alike.",
        "required": [
          "kcal",
          "protein_g",
          "fat_g",
          "carbs_g"
        ],
        "properties": {
          "carbs_g": {
            "type": "number",
            "format": "double",
            "description": "Carbohydrate, grams."
          },
          "fat_g": {
            "type": "number",
            "format": "double",
            "description": "Fat, grams."
          },
          "kcal": {
            "type": "number",
            "format": "double",
            "description": "Energy, kcal."
          },
          "protein_g": {
            "type": "number",
            "format": "double",
            "description": "Protein, grams."
          }
        }
      },
...
      "MealItemResponse": {
        "type": "object",
        "description": "One component of a meal.",
        "required": [
          "id",
          "position",
          "name",
          "grams",
          "grams_source",
          "kcal",
          "protein_g",
          "fat_g",
          "carbs_g",
          "macro_source"
        ],
        "properties": {
          "barcode": {
            "type": [
              "string",
              "null"
            ],
            "description": "EAN, when resolved from a barcode."
          },
          "carbs_g": {
            "type": "number",
            "format": "double",
            "description": "Carbohydrate, grams."
          },
          "confidence": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "0.0–1.0."
          },
          "fat_g": {
            "type": "number",
            "format": "double",
            "description": "Fat, grams."
          },
          "grams": {
            "type": "number",
            "format": "double",
            "description": "Edible mass in grams."
          },
          "grams_source": {
            "$ref": "#/components/schemas/GramsSource",
            "description": "Provenance of `grams`."
          },
          "id": {
            "type": "string",
            "description": "`meal_items.id`."
          },
          "kcal": {
            "type": "number",
            "format": "double",
            "description": "Energy."
          },
          "macro_source": {
            "$ref": "#/components/schemas/MacroSource",
            "description": "Provenance of the macros."
          },
          "name": {
            "type": "string",
            "description": "Component name."
          },
          "position": {
            "type": "integer",
            "format": "int32",
            "description": "Display order."
          },
          "protein_g": {
            "type": "number",
            "format": "double",
            "description": "Protein, grams."
          },
          "reasoning_note": {
            "type": [
              "string",
              "null"
            ],
            "description": "One line of \"why\", shown next to the item."
          }
        }
      },
      "MealResponse": {
        "type": "object",
        "description": "A meal with its items at the current revision.",
        "required": [
          "id",
          "client_uuid",
          "name_source",
          "revision",
          "status",
          "eaten_at",
          "timezone_offset",
          "portion_scale",
          "items",
          "totals",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "client_uuid": {
            "type": "string",
            "description": "Client-generated idempotency key."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Row creation time."
          },
          "dish_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Dish name as displayed."
          },
          "eaten_at": {
            "type": "string",
            "format": "date-time",
            "description": "When it was eaten (UTC)."
          },
          "error": {
            "type": [
              "string",
              "null"
            ],
            "description": "Failure reason when `status` is `failed` — visible, never silent (§5)."
          },
          "group_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "The sitting this meal belongs to, if any."
          },
          "group_size": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Client hint for how many photos the sitting will contain."
          },
          "id": {
            "type": "string",
            "description": "`meals.id`."
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MealItemResponse"
            },
            "description": "Items at `revision`, in display order."
          },
          "meal_type": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional meal-type label (\"breakfast\", \"snack\", …)."
          },
          "name_source": {
            "$ref": "#/components/schemas/NameSource",
            "description": "Which input path supplied the name — instrumented per §14.2."
          },
          "portion_scale": {
            "type": "number",
            "format": "double",
            "description": "\"Ate 60% of it\" multiplier; already applied to `items` and `totals`."
          },
          "revision": {
            "type": "integer",
            "format": "int32",
            "description": "Current revision."
          },
          "status": {
            "$ref": "#/components/schemas/MealStatus",
            "description": "Lifecycle status."
          },
          "thumbnail_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "Relative URL of the 768px thumbnail, or `None` for a manual entry."
          },
          "timezone_offset": {
            "type": "integer",
            "format": "int32",
            "description": "Minutes east of UTC at capture; buckets the local day."
          },
          "totals": {
            "$ref": "#/components/schemas/MacroTotals",
            "description": "Sum of `items`."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last modification time."
          },
          "user_comment": {
            "type": [
              "string",
              "null"
            ],
            "description": "The user's comment, when they typed one."
          }
        }
      },
      "MealStatus": {
        "type": "string",
        "description": "Lifecycle of a meal. `pending`/`analyzing` are in-flight, `needs_review`\nmeans the agent produced a draft the user has not confirmed, `confirmed`\nis the only status `recall_similar_meals` reads from, `failed` is loud.",
        "enum": [
          "Pending",
          "Analyzing",
          "NeedsReview",
          "Confirmed",
          "Failed"
        ]
      },
      "NameSource": {
        "type": "string",
        "description": "Which input path supplied the dish name. Instrumented in M4 to answer\nPRD §14.2 — whether the zero-keyboard paths actually get used.",
        "enum": [
          "Vision",
          "RecentChip",
          "ShareIntent",
          "Comment",
          "Manual"
        ]
      },
...
      "RecentDish": {
        "type": "object",
        "description": "One previously confirmed dish, ready to become a chip.",
        "required": [
          "dish_name",
          "dish_name_normalized",
          "last_eaten_at",
          "occurrences",
          "totals",
          "last_meal_id"
        ],
        "properties": {
          "dish_name": {
            "type": "string",
            "description": "The dish name as the user confirmed it."
          },
          "dish_name_normalized": {
            "type": "string",
            "description": "Normalized form — what recall matches on."
          },
          "last_eaten_at": {
            "type": "string",
            "format": "date-time",
            "description": "Most recent time this dish was eaten."
          },
          "last_meal_id": {
            "type": "string",
            "description": "The meal those totals came from, for a \"same as last time\" shortcut."
          },
          "occurrences": {
            "type": "integer",
            "format": "int64",
            "description": "How many times it has been confirmed. High counts are the flywheel\nworking."
          },
          "totals": {
            "$ref": "#/components/schemas/MacroTotals",
            "description": "Totals from the most recent confirmed occurrence, so the chip can show\n\"≈780 kcal\"."
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Session JWT. Web clients receive it as the HttpOnly `picweight_session` cookie after OIDC login; native clients obtain it from `POST /api/auth/token` and send it as `Authorization: Bearer`."
      }
    }
  },
...
}